refactor: update directory names
