+++ document/index.ts
