+++ document/useMethod.ts
