+++ document/useTimeAgo.ts
