+++ index.ts
@@ -7,7 +7,7 @@ export {
   useCacheUpdate,
 } from './cache';
 export { Straw } from './context';
-export { useDocumentList } from './document-list-resource';
-export { useDocument } from './document-resource';
-export { useList } from './list-resource';
+export { useDocument } from './document';
+export { useDocumentList } from './document-list';
+export { useList } from './list';
 export { StrawResource, useResource } from './resource';
